[GLT-4746] extracted Sample model to an interface
