+++ changes/add_sampleInterface.md
@@ -1,0 +1,1 @@
+Interface for sample data model
